requirements.txt
opencv-python
pyjwt
paho-mqtt
fastavro
kafka-python